requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
contourpy==1.3.0
cryptography==46.0.5
cycler==0.12.1
exceptiongroup==1.3.1
fonttools==4.60.2
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.47.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
langdetect==1.0.9
MarkupSafe==3.0.3
matplotlib==3.9.4
narwhals==2.18.0
nltk==3.9.2
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.6.0
proto-plus==1.27.1
protobuf==5.29.6
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rpds-py==0.27.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.3
streamlit==1.50.0
tenacity==9.1.2
textblob==0.19.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
wordcloud==1.9.6
zipp==3.23.0
